DT and Seg Constraints Done
