+++ constraint/src/main/java/gov/nist/hit/hl7/igamt/constraints/domain/assertion/Assertion.java
@@ -57,4 +57,8 @@ public abstract class Assertion {
   public void setScript(String script) {
     this.script = script;
   }
+
+  public String generateAssertionScript() {
+    return script;
+  }
 }
